Introduce data service and improve data coordination (#1006)
* Add signup tests and remove data sync on signup

Data cannot be sync'd with remote until the user has been authenticated, which happens upon login (not signup).

* Streamline strategy filters

No need to look up the session for every call to filter

* Make remote-store-sync strategy blocking

Ensures that remote transforms are sync'd with the store before the remote requests resolve.

* Make store-beforequery-remote-query and store-beforeupdate-remote-update strategies non-blocking

Don't block fulfillment of requests to the store with associated remote requests. Allows for optimistic UX.

* Introduce remote queryFail and updateFail strategies

* Hard fail when remote requests are attempted without auth credentials

Catch unauthenticated requests before they are made.

* Delay activation of remote source until coordinator has been activated

* Remove unnecessary transition after logOut

* Introduce data service

The data service manages activation, synchronization, and resetting of data. It ensure that there is one and only one color history palette. And it synchronizes local data with remote if there is no remote data.